Move State classes to inference.state package
