+++ src/main/java/com/example/inference/state/State.java
@@ -1,12 +1,9 @@
-package com.example.loader.weights;
+package com.example.inference.state;
 
-import com.example.core.model.tensor.ArrayFloatTensor;
 import com.example.core.model.tensor.FloatTensor;
 import com.example.model.Configuration;
 import uk.ac.manchester.tornado.api.types.arrays.FloatArray;
 import uk.ac.manchester.tornado.api.types.arrays.IntArray;
-
-import java.util.stream.Stream;
 
 /**
  * Base class for State
