+++ src/main/java/com/example/model/mistral/Mistral.java
@@ -3,8 +3,8 @@ package com.example.model.mistral;
 import com.example.inference.InferenceCore;
 import com.example.inference.InferenceEngine;
 import com.example.inference.sampler.Sampler;
-import com.example.loader.weights.LlamaState;
-import com.example.loader.weights.State;
+import com.example.inference.state.LlamaState;
+import com.example.inference.state.State;
 import com.example.loader.weights.Weights;
 import com.example.model.Model;
 import com.example.model.ModelType;
