+++ src/main/java/com/example/tornadovm/TornadoVMLayerPlanner.java
@@ -3,7 +3,7 @@ package com.example.tornadovm;
 import com.example.auxiliary.Tuple2;
 import com.example.model.Configuration;
 import com.example.model.Model;
-import com.example.loader.weights.State;
+import com.example.inference.state.State;
 import com.example.loader.weights.Weights;
 import uk.ac.manchester.tornado.api.GridScheduler;
 import uk.ac.manchester.tornado.api.ImmutableTaskGraph;
